hw: i386: move build_madt to acpi-build-common.c
Move build_madt() to acpi-build-common.c file, in order to be able to
compile acpi-build only for non HW reduced machine types (pc and q35).
